Module_1_Python/Lesson_12_Best_practices/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv
pyyaml==6.0.1
